+++ src/content-components/Image.tsx
@@ -35,6 +35,9 @@ export const Image: RendererConfig = {
     if (request.media.content?.type?.startsWith("image/")) {
       return RenderingPreference.NORMAL;
     }
+    if (request.media.content?.type?.startsWith("application/svg+xml")) {
+      return RenderingPreference.NORMAL;
+    }
     return RenderingPreference.INVALID;
   },
   render: ImageRenderer,